fix(judge-worker): 🔧 fix integer truncation, add retry, harden paths
Change memory_limit_mb from u64 to u32 to prevent silent truncation
(RW-H1). Add exponential backoff retry for result reporting (RW-H3).
Set workspace perms to 0o755 (SEC-M4). Use to_str() instead of
to_string_lossy for paths (RW-M4). Compare output as raw bytes
(RW-M2). Make seccomp profile path configurable via env (RW-M5).

+++ judge-worker-rs/src/types.rs
@@ -42,7 +42,7 @@ pub struct Submission {
     #[serde(rename = "timeLimitMs")]
     pub time_limit_ms: u64,
     #[serde(rename = "memoryLimitMb")]
-    pub memory_limit_mb: u64,
+    pub memory_limit_mb: u32,
     #[serde(rename = "testCases")]
     pub test_cases: Vec<TestCase>,
 }
